Removes utils library because it breaks ts-node https://github.com/TypeStrong/ts-node/issues/783.
Unfortunately, at the moment you have to choose whether to write imports that work in ts-node, or that work in the browser as es-modules.  Luckily, for this project there was one function in a second file so we can just pull that into the index file and circumvent the problem.  For other projects, resolving this issue will get a lot more messy.